uriManager ClientManager
