+++ src/HttpClient.php
@@ -10,13 +10,15 @@ namespace EasySwoole\HttpClient;
 
 
 use EasySwoole\HttpClient\Bean\Response;
-use EasySwoole\HttpClient\Bean\Url;
+use EasySwoole\HttpClient\Contract\ClientManager;
 use EasySwoole\HttpClient\Exception\InvalidUrl;
+use EasySwoole\HttpClient\Traits\UriManager;
 use Swoole\Coroutine\Http\Client;
 use Swoole\WebSocket\Frame;
 
 class HttpClient
 {
+    use UriManager;
     // HTTP 1.0/1.1 标准请求方法
     const METHOD_GET = 'GET';
     const METHOD_PUT = 'PUT';
@@ -37,16 +39,10 @@ class HttpClient
     const CONTENT_TYPE_X_WWW_FORM_URLENCODED = 'application/x-www-form-urlencoded';
 
     /**
-     * 当前请求路径
-     * @var Url
+     * 客户端管理
+     * @var ClientManager
      */
-    protected $url;
-
-    /**
-     * 协程客户端
-     * @var Client
-     */
-    protected $httpClient;
+    protected $clientManager;
 
     /**
      * 协程客户端设置项
@@ -101,41 +97,7 @@ class HttpClient
         $this->setConnectTimeout(5);
     }
 
-    public function setQuery(?array $data = null)
-    {
-        if($data){
-            $old = $this->url->getQuery();
-            parse_str($old,$old);
-            $this->url->setQuery(http_build_query($data + $old));
-        }
-        return $this;
-    }
-
     // --------  客户端配置设置方法  --------
-
-    /**
-     * 设置当前要请求的URL
-     * @param string $url 需要请求的网址
-     * @return HttpClient
-     * @throws InvalidUrl
-     */
-    public function setUrl($url): HttpClient
-    {
-        if ($url instanceof Url){
-            $this->url = $url;
-            return $this;
-        }
-        $info = parse_url($url);
-        if (empty($info['scheme'])) {
-            $info = parse_url('//' . $url); // 防止无scheme导致的host解析异常 默认作为http处理
-        }
-        $this->url = new Url($info);
-        if (empty($this->url->getHost())) {
-            throw new InvalidUrl("HttpClient: {$url} is invalid");
-        }
-        return $this;
-    }
-
 
     /**
      * 强制开启SSL
@@ -351,17 +313,18 @@ class HttpClient
 
     public function getClient(): Client
     {
-        if ($this->httpClient instanceof Client) {
+        if ($this->clientManager instanceof ClientManager) {
             $url = $this->parserUrlInfo();
-            $this->httpClient->host = $url->getHost();
-            $this->httpClient->port = $url->getPort();
-            $this->httpClient->ssl = $url->getIsSsl();
-            $this->httpClient->set($this->clientSetting);
-            return $this->httpClient;
+            $this->clientManager->getClient()->host = $url->getHost();
+            $this->clientManager->getClient()->port = $url->getPort();
+            $this->clientManager->getClient()->ssl = $url->getIsSsl();
+            $this->clientManager->getClient()->set($this->clientSetting);
+            return $this->clientManager->getClient();
         }
         $url = $this->parserUrlInfo();
-        $this->httpClient = new Client($url->getHost(), $url->getPort(), $url->getIsSsl());
-        $this->httpClient->set($this->clientSetting);
+        $this->clientManager = new \EasySwoole\HttpClient\Handler\Swoole\Client();
+        $this->clientManager->createClient($url->getHost(), $url->getPort(), $url->getIsSsl());
+        $this->clientManager->getClient()->set($this->clientSetting);
         return $this->getClient();
     }
 
@@ -369,57 +332,6 @@ class HttpClient
     {
         $this->getClient()->setMethod($method);
         return $this;
-    }
-
-    /**
-     * 解析当前的请求Url
-     * @throws InvalidUrl
-     */
-    protected function parserUrlInfo(?array $qurey = null)
-    {
-        // 请求时当前对象没有设置Url
-        if (!($this->url instanceof Url)) {
-            throw new InvalidUrl("HttpClient: Url is empty");
-        }
-
-        // 获取当前的请求参数
-        $path = $this->url->getPath();
-        $host = $this->url->getHost();
-        $port = $this->url->getPort();
-        $query = $this->url->getQuery();
-        $scheme = strtolower($this->url->getScheme());
-        if (empty($scheme)) {
-            $scheme = 'http';
-        }
-        // 支持的scheme
-        $allowSchemes = ['http' => 80, 'https' => 443, 'ws' => 80, 'wss' => 443];
-
-        // 只允许进行支持的请求
-        if (!array_key_exists($scheme, $allowSchemes)) {
-            throw new InvalidUrl("HttpClient: Clients are only allowed to initiate HTTP(WS) or HTTPS(WSS) requests");
-        }
-
-        // URL即使解析成功了也有可能存在HOST为空的情况
-        if (empty($host)) {
-            throw new InvalidUrl("HttpClient: Current URL is invalid because HOST is empty");
-        }
-
-        // 如果端口是空的 那么根据协议自动补全端口 否则使用原来的端口
-        if (empty($port)) {
-            $port = isset($allowSchemes[$scheme]) ? $allowSchemes[$scheme] : 80;
-            $this->url->setPort($port);
-        }
-
-        // 如果当前是443端口 或者enableSSL 则开启SSL安全链接
-        if ($this->enableSSL || $port === 443) {
-            $this->url->setIsSsl(true);
-        }
-
-        // 格式化路径和查询参数
-        $path = empty($path) ? '/' : $path;
-        $query = empty($query) ? '' : '?' . $query;
-        $this->url->setFullPath($path . $query);
-        return $this->url;
     }
 
     /**
@@ -834,9 +746,9 @@ class HttpClient
 
     function __destruct()
     {
-        if($this->httpClient instanceof Client){
-            if($this->httpClient->connected){
-                $this->httpClient->close();
+        if($this->clientManager->getClient() instanceof Client){
+            if($this->clientManager->getClient()){
+                $this->clientManager->getClient()->close();
             }
             $this->httpClient = null;
         }
